Refactor mult_submits functions to static methods
